requirements.txt
aiogram==3.23.0
python-dotenv==1.0.0
aiohttp>=3.9.0,<3.14
aiofiles>=23.2.1,<26
